Unify program modal handling
